Vagrantfile Perf Tweaks (Half Host CPU, 2GB RAM, Smart Drift Control, etc)

+++ shell/package.sh
@@ -2,19 +2,32 @@
 # Package up this vagrant box for distribution.
 # Usage: sudo bash package.sh
 
+echo "THIS WILL DESTROY AND RECREATE YOUR VAGRANT BOX."
+echo "It is NOT reccomended that you do this using your development environment."
+read -p "Are you sure you still want to do this? (y/n) " -n 1 -r
+echo
+if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+  echo "Cancelled."
+  exit
+fi
+
 BASEDIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
+
+cd "$BASEDIR/.."
 
 echo "Ensuring a clean machine"
 vagrant destroy --force
 
-cd "$BASEDIR/.."
+echo "Prepping config.rb and sites folder"
 cp config.rb-dist config.rb
+mkdir -p "$BASEDIR/../sites" >/dev/null 2>&1
 
-echo "Installing prerequirements"
+echo "Installing prerequisites"
 vagrant plugin install vagrant-vbguest
 vagrant plugin install vagrant-hostsupdater
 vagrant plugin install vagrant-useradd
 vagrant plugin install vagrant-bindfs
+vagrant plugin install vagrant-persistent-storage
 
 echo "Setting up Vagrant"
 sudo touch /etc/exports
@@ -22,6 +35,7 @@ vagrant up
 vagrant halt
 
 echo "Packaging Vagrant"
+rm -rf precip.box >/dev/null 2>&1
 vagrant package --base precip --output precip.box
 
 echo "Cleaning up"
